Revamp collision code
Rather than the almost completely disfunctional method used
previously (solely relying on the velocity of the object), this
attempts to provide a robust method of finding the current collision
side.

Other checks are added to try to ensure the correct side is
selected. This should be researched further.

Additionally, jumping can be done immediately after touching the
ground, which was useful in testing these changes.