Free player at bottom
This is similar to the commits which alter the camera when moving to
the left or right sides, except this is for the bottom. When falling,
the player moves from the center of the screen to the bottom when
reaching the bottom of the level. This removes the ugly green platform
from being visible below.

Because we want some ground still, the offset is used to shift this
effect upwards by a certain amount (in this case 50 pixels).

This works because the bottom platform starts at x = 0. If this
changes this code must be adjusted.