Allow smooth movement onto overlapping platforms
This is similar to the commit that fixed this for vertical platforms,
but for horizontal platforms instead. I realize that this is merely a
patch to deal with a bigger issue, but we'll get to it
eventually. This is a hobby project, and we haven't done a communist
revolution yet to ensure our free time is as productive as our work time.